+++ src/rpc/methods/eth.rs
@@ -2122,9 +2122,10 @@ impl RpcMethod<2> for EthGetCode {
             ..Default::default()
         };
 
+        let (state, _) = ctx.state_manager.tipset_state(&ts).await?;
         let api_invoc_result = 'invoc: {
             for ts in ts.chain_arc(ctx.store()) {
-                match ctx.state_manager.call(&message, Some(ts)) {
+                match ctx.state_manager.call_on_state(state, &message, Some(ts)) {
                     Ok(res) => {
                         break 'invoc res;
                     }
@@ -2173,10 +2174,8 @@ impl RpcMethod<3> for EthGetStorageAt {
             ResolveNullTipset::TakeOlder,
         )?;
         let to_address = FilecoinAddress::try_from(&eth_address)?;
-        let Some(actor) = ctx
-            .state_manager
-            .get_actor(&to_address, *ts.parent_state())?
-        else {
+        let (state, _) = ctx.state_manager.tipset_state(&ts).await?;
+        let Some(actor) = ctx.state_manager.get_actor(&to_address, state)? else {
             return Ok(make_empty_result());
         };
 
@@ -2195,7 +2194,7 @@ impl RpcMethod<3> for EthGetStorageAt {
         };
         let api_invoc_result = 'invoc: {
             for ts in ts.chain_arc(ctx.store()) {
-                match ctx.state_manager.call(&message, Some(ts)) {
+                match ctx.state_manager.call_on_state(state, &message, Some(ts)) {
                     Ok(res) => {
                         break 'invoc res;
                     }